[mc202:data_structs]
The reason I'm doing so is that I became convinced that it makes
harder for other developers to read my code. The code line gets way to
long when I'm doing the cast.

The reason I was using it before it is because for me it seems easier
to read my code but talking to other developers made me realize that
is not the case for every developer.